First step of converting to Wicked::Wizard
* Remove our state management for steps
* Make our existing specs pass for our existing routes
* Use only #steps and #render_wizard and #next_step from Wicked::Wizard
* Define the steps in the controller and rename our partials, stop
  managing that state ourselves
