+++ app/helpers/research_sessions_helper.rb
@@ -1,0 +1,2 @@
+module ResearchSessionsHelper
+end
