+++ app/models/research_session.rb
@@ -1,0 +1,3 @@
+class ResearchSession
+  PARAM_KEYS = [:age, :participant_name, :guardian_name]
+end
