+++ app/views/research_sessions/index.html.erb
@@ -5,5 +5,5 @@
   with research participants.
 </p>
 <p class="button-bar">
-  <%= link_to 'Create new form', {:controller => :questions, :action => :start}, class: 'button'  %>
+  <%= link_to 'Create new form', {:controller => :research_sessions, :action => :start}, class: 'button'  %>
 </p>
