+++ app/views/research_sessions/methods.html.erb
