+++ app/views/research_sessions/name.html.erb
